+++ grant-contacts/src/commonTest/kotlin/dev/brewkits/grant/contacts/system/ContactsSystemTest.kt
@@ -4,7 +4,7 @@ import dev.brewkits.grant.AppGrant
 import dev.brewkits.grant.GrantGroupHandler
 import dev.brewkits.grant.GrantHandler
 import dev.brewkits.grant.GrantStatus
-import dev.brewkits.grant.contacts.fakes.FakeGrantManager
+import dev.brewkits.grant.testing.FakeGrantManager
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.TestScope