+++ grant-core/src/commonTest/kotlin/dev/brewkits/grant/regression/Issue32MutexDoubleUnlockTest.kt
@@ -3,7 +3,7 @@ package dev.brewkits.grant.regression
 import dev.brewkits.grant.AppGrant
 import dev.brewkits.grant.GrantHandler
 import dev.brewkits.grant.GrantStatus
-import dev.brewkits.grant.fakes.FakeGrantManager
+import dev.brewkits.grant.testing.FakeGrantManager
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.advanceUntilIdle